pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project.urls]
homepage = "https://github.com/MolSSI/QCEngine"
changelog = "https://github.com/MolSSI/QCEngine/blob/master/docs/source/changelog.rst"
documentation = "https://molssi.github.io/QCEngine/"
issues = "https://github.com/MolSSI/QCEngine/issues"

[project]
name = "qcengine"
#version = "0.31.0"
dynamic = ["version"]
requires-python = ">=3.9"
description = "A compute wrapper for Quantum Chemistry, ingesting and producing QCSchema for a variety of QC programs."
authors = [
  { name="The QCArchive Development Team", email="qcarchive@molssi.org" },
]
license = { file="LICENSE" }
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Framework :: Pydantic",
    "Framework :: Pydantic :: 2",
]

dependencies = [
    "pyyaml",
    "py-cpuinfo",
    "psutil",
    "qcelemental>=0.50a1,<0.70.0",
    "pydantic >=2.1",
    "pydantic-settings",
    "packaging",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "codecov",
]
lint = [
    "pre-commit",
    "black >=22.1.0,<23.0a0",  # if running outside of pre-commit
    "isort >=5.13.2",  # if running outside of pre-commit
    # "mypy",
    # "autoflake",
    # "flake8",
]
docs = [  # probably >=py39
    "numpydoc",
    # "docutils",
    "sphinx >=7.0.0",
    "sphinxcontrib-napoleon",
    "sphinx-rtd-theme",
    "autodoc-pydantic>=2.0",
    # "sphinx-automodapi",
    # "sphinx-autodoc-typehints",
    # "pydantic ==2.5.0",  # for now
    # "pydantic-settings ==2.2.0",  # for now
]

[project.scripts]
qcengine = "qcengine.cli:main"

[tool.black]
line-length = 120
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 120

[tool.setuptools]
packages = ["qcengine"]

[aliases]
test="pytest"

[tool.setuptools.package-data]

[tool.setuptools_scm]

[tool.versioningit.vcs]
default-tag = "0.0.99" # useful for CI/shallow clones

[tool.pytest.ini_options]
#filterwarnings = [
#    "ignore::DeprecationWarning",
#    "ignore::PendingDeprecationWarning",
#]
markers = [
    "long",
    """slow: marks tests as slow (deselect with '-m "not slow"')""",
    "smoke",
]
